requirements.txt
flask==1.0.3
Pillow==6.0.0
